+++ plugins/auth-backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-auth-backend
 
+## 0.5.1
+
+### Patch Changes
+
+- 699c2e9ddc: export minimal typescript types for OIDC provider
+- Updated dependencies
+  - @backstage/backend-common@0.9.14
+  - @backstage/catalog-model@0.9.8
+
 ## 0.5.0
 
 ### Minor Changes